refactor: Move assets Vite plugin into src
