+++ tsconfig.node.json
@@ -5,5 +5,5 @@
     "moduleResolution": "Node",
     "allowSyntheticDefaultImports": true
   },
-  "include": ["vite.config.ts", "lib/assets-plugin.ts", "src/types.ts"]
+  "include": ["vite.config.ts", "src/lib/assets-plugin.ts"]
 }
